+++ R/SimesThresholdFamily.R
@@ -19,23 +19,7 @@
 ##' sk <- SimesThresholdFamily(12)
 ##' thr <- sk(0.2)
 ##'
-SimesThresholdFamily <- structure(function(
-### Simes' threshold family
-    m,
-### The number of hypotheses tested
-    kMax=m
-### For simultaneous control of (\eqn{k}-FWER for all \eqn{k \le
-### k[max]}).
-    )
-   {
-       sk <- function(alpha) qnorm(1-min(alpha, 1)*(1:kMax)/m)
-       return(sk)
-### A threshold function (on the scale of test statistics) based on
-### the classical family of thresholds introduced by Simes (1986):
-### \eqn{\alpha*k/m}. This family yields joint FWER control at level
-### (at most) \eqn{\alpha} if the test statistics are positively
-### dependent (PRDS) under H0.
-   }, ex=function(){
-       sk <- SimesThresholdFamily(12)
-       thr <- sk(0.2)
-   })
+SimesThresholdFamily <- function(m, kMax=m){
+    sk <- function(alpha) qnorm(1-min(alpha, 1)*(1:kMax)/m)
+    return(sk)
+}