+++ src/coverage.cpp
@@ -7,8 +7,9 @@ using namespace Rcpp;
 //'
 //' @param thr A numeric vector of length \code{m}, a threshold family. Should
 //'   be sorted in decreasing order.
-//' @param Z A \code{m} x \code{B} matrix of \code{B} Monte-Carlo samples of test statistics
-//'   under the null hypothesis. Each column should be sorted in decreasing order.
+//' @param Z A \code{m} x \code{B} matrix of \code{B} Monte-Carlo
+//'    samples of *ordered* test statistics under the null
+//'    hypothesis. Each column should be sorted in decreasing order.
 //' @return The empirical coverage of \code{thr} according to \code{Z}.
 //'
 //' @details The empirical coverage of \code{thr} according to
@@ -17,9 +18,8 @@ using namespace Rcpp;
 //'   the test statistics under the null hypothesis exceeds the \eqn{k}-th
 //'   largest value of \code{thr}, that is, \code{thr[k]}.
 //'
-//' @export
 // [[Rcpp::export]]
-NumericVector coverage(NumericVector thr, arma::mat Z) {
+NumericVector empiricalCoverageO(NumericVector thr, arma::mat Z) {
     // Calculate P_n(k-max(thr) > thr(k))
     int m = thr.size();
     int B = Z.n_cols;
@@ -45,7 +45,7 @@ m <- 10
 B <- 20
 thr <- sort(rnorm(m), decreasing=TRUE)
 Z <- matrix(rnorm(m*B), m, B)
-p <- coverage(thr, Z)
+p <- empiricalCoverageO(thr, Z)
 
 ## R version:
 isAbove <- sweep(Z, 1, thr,  ">")